requirements.txt
tasklib
icalendar
flask
gevent
pyyaml
caldav
pytest-mock

# separate these
pytest
pytest-cov
